+++ scripts/functions-deploy-tests/run.sh
@@ -2,7 +2,7 @@
 set -e # Immediately exit on failure
 
 # Globally link the CLI for the testing framework
-./scripts/npm-link.sh
+./scripts/clean-install.sh
 
 # Create a secret for testing if it doesn't exist
 firebase functions:secrets:get TOP --project $GCLOUD_PROJECT || (echo secret | firebase functions:secrets:set --data-file=- TOP --project $GCLOUD_PROJECT -f)