+++ src/commands/ext-install.ts
@@ -11,10 +11,9 @@ import { getProjectId, needProjectId } from "../projectUtils";
 import * as extensionsApi from "../extensions/extensionsApi";
 import { ExtensionVersion, ExtensionSource } from "../extensions/types";
 import * as refs from "../extensions/refs";
-import { displayWarningPrompts } from "../extensions/warnings";
+import * as secretsUtils from "../extensions/secretsUtils";
 import * as paramHelper from "../extensions/paramHelper";
 import {
-  confirm,
   createSourceFromLocation,
   ensureExtensionsApiEnabled,
   logPrefix,
@@ -29,9 +28,10 @@ import { getRandomString } from "../extensions/utils";
 import { requirePermissions } from "../requirePermissions";
 import * as utils from "../utils";
 import { track } from "../track";
-import * as experiments from "../experiments";
+import { confirm } from "../prompt";
 import { Options } from "../options";
 import * as manifest from "../extensions/manifest";
+import { displayDeveloperTOSWarning } from "../extensions/tos";
 
 marked.setOptions({
   renderer: new TerminalRenderer(),
@@ -42,11 +42,8 @@ marked.setOptions({
  */
 export const command = new Command("ext:install [extensionName]")
   .description(
-    "install an official extension if [extensionName] or [extensionName@version] is provided; " +
-      (experiments.isEnabled("extdev")
-        ? "install a local extension if [localPathOrUrl] or [url#root] is provided; install a published extension (not authored by Firebase) if [publisherId/extensionId] is provided "
-        : "") +
-      "or run with `-i` to see all available extensions."
+    "add an uploaded extension to firebase.json if [publisherId/extensionId] is provided;" +
+      "or, add a local extension if [localPath] is provided"
   )
   .option("--local", "deprecated")
   .withForce()
@@ -165,13 +162,7 @@ async function infoExtensionVersion(args: {
   extensionVersion: ExtensionVersion;
 }): Promise<void> {
   const ref = refs.parse(args.extensionName);
-  const extension = await extensionsApi.getExtension(refs.toExtensionRef(ref));
   await displayExtInfo(args.extensionName, ref.publisherId, args.extensionVersion.spec, true);
-  await displayWarningPrompts(
-    ref.publisherId,
-    extension.registryLaunchStage,
-    args.extensionVersion
-  );
 }
 
 interface InstallExtensionOptions {
@@ -202,6 +193,10 @@ async function installToManifest(options: InstallExtensionOptions): Promise<void
     );
   }
 
+  if (secretsUtils.usesSecrets(spec)) {
+    await secretsUtils.ensureSecretManagerApiEnabled(options);
+  }
+
   const config = manifest.loadConfig(options);
 
   let instanceId = spec.name;
@@ -211,7 +206,7 @@ async function installToManifest(options: InstallExtensionOptions): Promise<void
 
   const paramBindingOptions = await paramHelper.getParams({
     projectId,
-    paramSpecs: spec.params.concat(spec.systemParams ?? []),
+    paramSpecs: (spec.params ?? []).concat(spec.systemParams ?? []),
     nonInteractive,
     paramsEnvPath,
     instanceId,
@@ -243,5 +238,5 @@ async function installToManifest(options: InstallExtensionOptions): Promise<void
     config,
     { nonInteractive, force: force ?? false }
   );
-  manifest.showPostDeprecationNotice();
+  displayDeveloperTOSWarning();
 }