+++ src/frameworks/nuxt/interfaces.ts
@@ -1,5 +1,8 @@
-export interface NuxtDependency {
-  version?: string;
-  resolved?: string;
-  overridden?: boolean;
+// TODO: define more fields as needed
+// The NuxtOptions interface is huge and depends on multiple external types
+// and packages. For now only the fields that are being used are defined.
+export interface NuxtOptions {
+  dir: {
+    public: string;
+  };
 }