+++ src/test/gcp/cloudfunctionsv2.spec.ts
@@ -23,6 +23,7 @@ describe("cloudfunctionsv2", () => {
     entryPoint: "function",
     runtime: "nodejs16",
     codebase: projectConfig.DEFAULT_CODEBASE,
+    runServiceId: "service",
   };
 
   const CLOUD_FUNCTION_V2_SOURCE: cloudfunctionsv2.StorageSource = {
@@ -31,26 +32,26 @@ describe("cloudfunctionsv2", () => {
     generation: 42,
   };
 
-  const CLOUD_FUNCTION_V2: Omit<cloudfunctionsv2.CloudFunction, cloudfunctionsv2.OutputOnlyFields> =
-    {
-      name: "projects/project/locations/region/functions/id",
-      buildConfig: {
-        entryPoint: "function",
-        runtime: "nodejs16",
-        source: {
-          storageSource: CLOUD_FUNCTION_V2_SOURCE,
-        },
-        environmentVariables: {},
+  const CLOUD_FUNCTION_V2: cloudfunctionsv2.InputCloudFunction = {
+    name: "projects/project/locations/region/functions/id",
+    buildConfig: {
+      entryPoint: "function",
+      runtime: "nodejs16",
+      source: {
+        storageSource: CLOUD_FUNCTION_V2_SOURCE,
       },
-      serviceConfig: {
-        availableMemory: `${backend.DEFAULT_MEMORY}Mi`,
-      },
-    };
+      environmentVariables: {},
+    },
+    serviceConfig: {
+      availableMemory: `${backend.DEFAULT_MEMORY}Mi`,
+    },
+  };
 
   const RUN_URI = "https://id-nonce-region-project.run.app";
-  const HAVE_CLOUD_FUNCTION_V2: cloudfunctionsv2.CloudFunction = {
+  const HAVE_CLOUD_FUNCTION_V2: cloudfunctionsv2.OutputCloudFunction = {
     ...CLOUD_FUNCTION_V2,
     serviceConfig: {
+      service: "service",
       uri: RUN_URI,
     },
     state: "ACTIVE",
@@ -116,10 +117,7 @@ describe("cloudfunctionsv2", () => {
           channel: "projects/myproject/locations/us-wildwest11/channels/mychannel",
         },
       };
-      const eventGcfFunction: Omit<
-        cloudfunctionsv2.CloudFunction,
-        cloudfunctionsv2.OutputOnlyFields
-      > = {
+      const eventGcfFunction: cloudfunctionsv2.InputCloudFunction = {
         ...CLOUD_FUNCTION_V2,
         eventTrigger: {
           eventType: "google.cloud.audit.log.v1.written",
@@ -266,10 +264,7 @@ describe("cloudfunctionsv2", () => {
         ],
       };
 
-      const fullGcfFunction: Omit<
-        cloudfunctionsv2.CloudFunction,
-        cloudfunctionsv2.OutputOnlyFields
-      > = {
+      const fullGcfFunction: cloudfunctionsv2.InputCloudFunction = {
         ...CLOUD_FUNCTION_V2,
         labels: {
           ...CLOUD_FUNCTION_V2.labels,
@@ -317,10 +312,7 @@ describe("cloudfunctionsv2", () => {
         availableMemoryMb: 128,
       };
 
-      const complexGcfFunction: Omit<
-        cloudfunctionsv2.CloudFunction,
-        cloudfunctionsv2.OutputOnlyFields
-      > = {
+      const complexGcfFunction: cloudfunctionsv2.InputCloudFunction = {
         ...CLOUD_FUNCTION_V2,
         eventTrigger: {
           eventType: events.v2.PUBSUB_PUBLISH_EVENT,
@@ -355,7 +347,7 @@ describe("cloudfunctionsv2", () => {
         concurrency: 40,
         cpu: 2,
       };
-      const gcfFunction: Omit<cloudfunctionsv2.CloudFunction, cloudfunctionsv2.OutputOnlyFields> = {
+      const gcfFunction: cloudfunctionsv2.InputCloudFunction = {
         ...CLOUD_FUNCTION_V2,
         serviceConfig: {
           ...CLOUD_FUNCTION_V2.serviceConfig,
@@ -404,7 +396,7 @@ describe("cloudfunctionsv2", () => {
     });
 
     it("should copy run service IDs", () => {
-      const fn: cloudfunctionsv2.CloudFunction = {
+      const fn: cloudfunctionsv2.OutputCloudFunction = {
         ...HAVE_CLOUD_FUNCTION_V2,
         serviceConfig: {
           ...HAVE_CLOUD_FUNCTION_V2.serviceConfig,
@@ -700,6 +692,21 @@ describe("cloudfunctionsv2", () => {
         codebase: "my-codebase",
         hash: "my-hash",
       });
+    });
+
+    it("should convert function without serviceConfig", () => {
+      const expectedEndpoint = {
+        ...ENDPOINT,
+        platform: "gcfv2",
+        httpsTrigger: {},
+      };
+      delete expectedEndpoint.runServiceId;
+      expect(
+        cloudfunctionsv2.endpointFromFunction({
+          ...HAVE_CLOUD_FUNCTION_V2,
+          serviceConfig: undefined,
+        })
+      ).to.deep.equal(expectedEndpoint);
     });
   });