+++ CHANGELOG.md
@@ -1,2 +1,3 @@
-- Releases Cloud Firestore emulator v1.16.1, which adds support for read_time in ListCollectionIds.
+- Fixes an issue in the EventArc emualtor where events missing optional fields would cause crashes. (#5803)
+- Fixes an issue running `firebase emulators:start` and `firebase deploy` when Python Cloud Functions directory path has spaces. (#5830)
 - Supports parametrized configuration of ServiceAccount in Functions and Function triggers.